+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/Customization.java
@@ -1,14 +1,14 @@
 package com.soramusoka.destinyApiClient.dto_layer.account_summary;
 
 public class Customization {
-    public double personality;
-    public double face;
-    public double skinColor;
-    public double lipColor;
-    public double eyeColor;
-    public double hairColor;
-    public double featureColor;
-    public double decalColor;
+    public long personality;
+    public long face;
+    public long skinColor;
+    public long lipColor;
+    public long eyeColor;
+    public long hairColor;
+    public long featureColor;
+    public long decalColor;
     public boolean wearHelmet;
     public int hairIndex;
     public int featureIndex;
