+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/Equipment.java
@@ -1,6 +1,6 @@
 package com.soramusoka.destinyApiClient.dto_layer.account_summary;
 
 public class Equipment {
-    public double itemHash;
+    public long itemHash;
     public Dyes[] dyes;
 }
