+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/Arrangement.java
@@ -1,6 +1,6 @@
 package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 
 public class Arrangement {
-    public double classHash;
+    public long classHash;
     public int gearArtArrangementIndex;
 }
