+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/BucketDefinition.java
@@ -3,7 +3,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class BucketDefinition extends BaseDefinition {
-    public double bucketHash;
+    public long bucketHash;
     public String bucketName;
     public String bucketDescription;
     public int scope;
