+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/CustomDyeExpressionStep.java
@@ -2,7 +2,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 
 public class CustomDyeExpressionStep {
     public int stepOperator;
-    public double flagHash;
+    public long flagHash;
     public int valueHash;
     public int value;
 }
