+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/DamageTypeDefinition.java
@@ -3,7 +3,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class DamageTypeDefinition extends BaseDefinition {
-    public double damageTypeHash;
+    public long damageTypeHash;
     public String identifier;
     public String damageTypeName;
     public String iconPath;
