+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/DestinationDefinition.java
@@ -3,11 +3,11 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class DestinationDefinition extends BaseDefinition {
-    public double destinationHash;
+    public long destinationHash;
     public String destinationName;
     public String destinationDescription;
     public String icon;
-    public double placeHash;
+    public long placeHash;
     public String destinationIdentifier;
     public String locationIdentifier;
 }
