+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/EquippingBlock.java
@@ -9,6 +9,6 @@ public class EquippingBlock {
     public Object[] lockedDyes;
     public Object[] customDyes;
     public CustomDyeExpression customDyeExpression;
-    public double weaponPatternHash;
+    public long weaponPatternHash;
     public Arrangement[] arrangements;
 }
