+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/ItemCategoryDefinition.java
@@ -3,7 +3,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class ItemCategoryDefinition extends BaseDefinition {
-    public double itemCategoryHash;
+    public long itemCategoryHash;
     public String identifier;
     public boolean visible;
     public String title;
