+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/RaceDefinition.java
@@ -3,7 +3,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class RaceDefinition extends BaseDefinition {
-    public double raceHash;
+    public long raceHash;
     public int raceType;
     public String raceName;
     public String raceNameMale;
