+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/SourceDefinition.java
@@ -3,7 +3,7 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class SourceDefinition extends BaseDefinition {
-    public double sourceHash;
+    public long sourceHash;
     public int category;
     public String identifier;
     public String sourceName;
