+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/definitions/TalentGridDefinition.java
@@ -3,10 +3,10 @@ package com.soramusoka.destinyApiClient.dto_layer.account_summary.definitions;
 import com.soramusoka.destinyApiClient.dto_layer.common.BaseDefinition;
 
 public class TalentGridDefinition extends BaseDefinition {
-    public double gridHash;
+    public long gridHash;
     public int maxGridLevel;
     public int gridLevelPerColumn;
-    public double progressionHash;
+    public long progressionHash;
     public TalentGridNode[] nodes;
     public int calcMaxGridLevel;
     public int calcProgressToMaxLevel;
