+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/aggregate_activity_stats/AggregateActivityStat.java
@@ -1,6 +1,6 @@
 package com.soramusoka.destinyApiClient.dto_layer.aggregate_activity_stats;
 
 public class AggregateActivityStat {
-    public double activityHash;
+    public long activityHash;
     public AggregateActivityValueGroup values;
 }
