+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/common/BaseDefinition.java
@@ -1,6 +1,6 @@
 package com.soramusoka.destinyApiClient.dto_layer.common;
 
 public class BaseDefinition {
-    public double hash;
+    public long hash;
     public int index;
 }
