+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/common/BaseInventoryItem.java
@@ -1,9 +1,9 @@
 package com.soramusoka.destinyApiClient.dto_layer.common;
 
 public class BaseInventoryItem {
-    public double itemHash;
+    public long itemHash;
     public int damageType;
-    public double damageTypeHash;
+    public long damageTypeHash;
     public boolean isGridComplete;
     public int transferStatus;
     public int state;
