+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/inventory_item/UpgradeMaterial.java
@@ -1,7 +1,7 @@
 package com.soramusoka.destinyApiClient.dto_layer.inventory_item;
 
 public class UpgradeMaterial {
-    public double itemHash;
+    public long itemHash;
     public boolean deleteOnAction;
     public int count;
     public boolean omitFromRequirements;
